+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Features
 - [#1261](https://github.com/crypto-org-chain/chain-main/pull/1261) Implement tokenomics (inflation decay, base rewards topup and tiered rewards module)
+- [#1265](https://github.com/crypto-org-chain/chain-main/pull/1265) feat(x/tieredrewards): add MsgExitTierWithDelegation - allow position to exit tier by transferring delegation back to owner directly.
 
 ### Improvements
 - [#1269](https://github.com/crypto-org-chain/chain-main/pull/1269) feat: MsgClaimTierRewards now accepts multiple position IDs to claim rewards in a single transaction.